Added semi-working personal messaging system
Added a personal messaging system. Messages are written to the server, but cant be read yet. The system is originally in a new module personal-messaging module, but in attemt to fix some issues, the module is skipped for now and the components are direcly hooked to the app module.
Made some changes to webService